requirements.txt
flask==2.3.3
flask_socketio==5.3.6
python-dotenv==1.0.0